initial updating steps
Attempting to containerize this to make managing dependencies much
simpler, but running bpftrace inside a container is difficult, so
pushing this current attempt in order to get feedback and experiment
elsewhere.

The scripts and programs from the previous sprint are copied here with
some modifications. The program do_it_all.c is now ntck.bash.
send_raw_eth_2.c is now sendeth.c, and it has been updated to accept the
network interface to use as a command-line argument. Some redundant
lines in plot_results.py are also removed.

The script buildnrun.sh builds and runs the Docker container.
As mentioned, the Docker container currently does not work.